+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/room/read/DefaultReadService.kt
@@ -85,6 +85,15 @@ internal class DefaultReadService @AssistedInject constructor(
         return isEventRead(monarchy, userId, roomId, eventId)
     }
 
+    override fun getReadMarker(): Optional<String> {
+        var res: String? = null
+        monarchy.doWithRealm { realm ->
+            val readMarkerEntity = ReadMarkerEntity.where(realm, roomId = roomId).findFirst()
+            res = readMarkerEntity?.eventId
+        }
+        return Optional(res)
+    }
+
     override fun getReadMarkerLive(): LiveData<Optional<String>> {
         val liveRealmData = monarchy.findAllMappedWithChanges(
                 { ReadMarkerEntity.where(it, roomId) },